+++ src/index.ts
@@ -1,40 +1,121 @@
-import call from './RequestHandler';
-import { Cumulonimbus as __Cumulonimbus } from './Cumulonimbus';
-import toFormData from './DataToFormData';
+import fetch from "isomorphic-fetch";
+import FormData from "isomorphic-form-data";
 
-const WITH_BODY = { 'Content-Type': 'application/json' };
+// Get version from package.json
+const version = require("./package.json").version;
 
-export const Cumulonimbus = __Cumulonimbus;
+// deep merge two objects without overwriting existing properties
+function merge(obj1: any, obj2: any) {
+  const result = { ...obj1 };
+  for (const key in obj2) {
+    if (obj2.hasOwnProperty(key)) {
+      if (typeof obj2[key] === "object") {
+        result[key] = merge(result[key], obj2[key]);
+      } else {
+        result[key] = obj2[key];
+      }
+    }
+  }
+  return result;
+}
 
-export class Client {
+class Cumulonimbus {
   private token: string;
+  private options: Cumulonimbus.ClientOptions;
 
-  private options: __Cumulonimbus.ClientOptions;
+  constructor(token: string, options?: Cumulonimbus.ClientOptions) {
+    this.token = token;
+    this.options = options || {};
+  }
 
-  /// why spend 6 minutes implementing everything by hand when we can spend 6 hours trying to automate it
+  private async call<T>(
+    url: string,
+    options: Cumulonimbus.APICallRequestInit = {}
+  ): Promise<Cumulonimbus.APIResponse<T>> {
+    const opts = merge(options, {
+      headers: {
+        "User-Agent": `Cumulonimbus/${version}`,
+      },
+    });
+
+    const res = await fetch(
+      (options.baseURL || this.options.baseURL) + url,
+      opts
+    );
+
+    let ratelimit: Cumulonimbus.RatelimitData | null = null;
+
+    if (res.headers.get("X-Ratelimit-Limit")) {
+      ratelimit = {
+        max: Number(res.headers.get("X-RateLimit-Limit")),
+        remaining: Number(res.headers.get("X-RateLimit-Remaining")),
+        reset: Number(res.headers.get("X-RateLimit-Reset")),
+      };
+    }
+
+    // Check if the response is a 413, and if so, construct a new BODY_TOO_LARGE error, as we can't be sure the server returns a proper error
+    if (res.status === 413) {
+      throw new Cumulonimbus.ResponseError(
+        {
+          code: "BODY_TOO_LARGE_ERROR",
+          message: "Body Too Large",
+        },
+        ratelimit
+      );
+    }
+
+    try {
+      let json = await res.json();
+      if (!res.ok) throw new Cumulonimbus.ResponseError(json, ratelimit);
+      else return { result: json, ratelimit, response: res };
+    } catch (error) {
+      if (error instanceof Cumulonimbus.ResponseError) throw error;
+      else
+        throw new Cumulonimbus.ResponseError(
+          {
+            code: "GENERIC_ERROR",
+            message: null,
+          },
+          ratelimit
+        );
+    }
+  }
+
+  private async authenticatedCall<T>(
+    url: string,
+    options: Cumulonimbus.APICallRequestInit = {}
+  ) {
+    const opts = merge(options, {
+      headers: {
+        Authorization: `Bearer ${this.token}`,
+      },
+    });
+    return this.call<T>(url, opts);
+  }
+
   private manufactureMethod<T extends any[], M>(
     endpointTemplate: string | ((...args: T) => string),
     method: string,
-    headers: { [key: string]: string },
-    bodyTemplate: string | null | ((...args: T) => any)
-  ): (...args: T) => Promise<M> {
-    return async (...args: T): Promise<M> => {
+    headers: { [key: string]: string } = {},
+    bodyTemplate: string | null | ((...args: T) => string) = null
+  ): (...args: T) => Promise<Cumulonimbus.APIResponse<M>> {
+    return async (...args: T): Promise<Cumulonimbus.APIResponse<M>> => {
       try {
         let endpoint = (
-          typeof endpointTemplate === 'string'
+          typeof endpointTemplate === "string"
             ? () => endpointTemplate
             : endpointTemplate
         )(...args);
         let res = await this.authenticatedCall<M>(endpoint, {
           method,
           headers,
-          body: (typeof bodyTemplate === 'function'
+          body: (typeof bodyTemplate === "function"
             ? bodyTemplate
-            : () => bodyTemplate)(...args)
+            : () => bodyTemplate)(...args),
         });
 
-        if (res.res.ok) return res.payload;
-        else throw new Error('https://youtu.be/snKJPEVbQoE?t=20');
+        if (res.response.ok) return res;
+        else throw new Error("https://youtu.be/snKJPEVbQoE?t=20");
       } catch (error) {
         throw error;
       }
@@ -42,456 +123,240 @@ export class Client {
   }
 
   private manufactureMethodGet<T extends any[], M>(
-    endpointTemplate: string | ((...args: T) => string)
-  ): (this: Client, ...args: T) => Promise<M> {
-    return this.manufactureMethod<T, M>(endpointTemplate, 'GET', {}, null);
+    endpointTemplate: string | ((...args: T) => string),
+    headers: { [key: string]: string } = {}
+  ): (...args: T) => Promise<Cumulonimbus.APIResponse<M>> {
+    return this.manufactureMethod<T, M>(endpointTemplate, "GET", headers, null);
   }
 
   public static async login(
     user: string,
     pass: string,
     rememberMe: boolean = false,
-    options?: __Cumulonimbus.ClientOptions
-  ): Promise<Client> {
-    try {
-      let res = await call<__Cumulonimbus.Data.SuccessfulAuth>(
-        '/user/session',
-        {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json'
-          },
-          body: JSON.stringify({ user, pass, rememberMe }),
-          baseURL: options && options.baseURL ? options.baseURL : undefined
-        }
-      );
-
-      if (res.res.status === 201) {
-        let rtn = new Client(res.payload.token, options);
-        return rtn;
-      } else {
-        throw new Error('https://youtu.be/snKJPEVbQoE?t=20');
+    options?: Cumulonimbus.ClientOptions
+  ): Promise<Cumulonimbus> {
+    const res = await fetch(
+      options && options.baseURL ? options.baseURL : Cumulonimbus.BASE_URL,
+      {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+          "User-Agent": `Cumulonimbus/${version}`,
+        },
+        body: JSON.stringify({
+          user,
+          pass,
+          rememberMe,
+        }),
       }
-    } catch (error) {
-      throw error;
+    );
+
+    let ratelimit: Cumulonimbus.RatelimitData | null = null;
+
+    if (res.headers.get("X-Ratelimit-Limit")) {
+      ratelimit = {
+        max: Number(res.headers.get("X-RateLimit-Limit")),
+        remaining: Number(res.headers.get("X-RateLimit-Remaining")),
+        reset: Number(res.headers.get("X-RateLimit-Reset")),
+      };
     }
+
+    const json = await res.json();
+
+    if (!res.ok) throw new Cumulonimbus.ResponseError(json, ratelimit);
+
+    return new Cumulonimbus(json.token, options);
   }
 
-  public static async createAccount(
+  public static async register(
     username: string,
-    password: string,
-    repeatPassword: string,
     email: string,
+    password: string,
+    confirmPassword: string,
     rememberMe: boolean = false,
-    options?: __Cumulonimbus.ClientOptions
-  ) {
-    try {
-      let res = await call<__Cumulonimbus.Data.SuccessfulAuth>('/user', {
-        method: 'POST',
+    options?: Cumulonimbus.ClientOptions
+  ): Promise<Cumulonimbus> {
+    const res = await fetch(
+      options && options.baseURL ? options.baseURL : Cumulonimbus.BASE_URL,
+      {
+        method: "POST",
         headers: {
-          'Content-Type': 'application/json'
+          "Content-Type": "application/json",
+          "User-Agent": `Cumulonimbus/${version}`,
         },
         body: JSON.stringify({
           username,
-          password,
           email,
-          repeatPassword,
-          rememberMe
+          password,
+          rememberMe,
+          repeatPassword: confirmPassword,
         }),
-        baseURL: options && options.baseURL ? options.baseURL : undefined
-      });
-
-      if (res.res.ok) {
-        let rtn = new Client(res.payload.token, options);
-        return rtn;
-      } else throw new Error('https://youtu.be/snKJPEVbQoE?t=20');
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  public constructor(token: string, options?: __Cumulonimbus.ClientOptions) {
-    this.options = options || {};
-    this.token = token;
-  }
-
-  private async authenticatedCall<T>(
-    url: string,
-    options: RequestInit = { headers: {} }
-  ): Promise<{ res: Response; payload: T }> {
-    let opts = Object.assign(Object.create(null), options);
-    if (!opts.headers) opts.headers = {};
-    opts.baseURL = this.options.baseURL;
-    (opts.headers as any).Authorization = this.token;
-    try {
-      let authenticatedCall = await call<T>(url, opts);
-      return authenticatedCall;
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  public sanityCheck = this.manufactureMethodGet<
-    [],
-    __Cumulonimbus.Data.SanityCheck
-  >('/');
-
-  public thumbnailSanityCheck = async () => {
-    let res = await call<__Cumulonimbus.Data.SanityCheck>('/', {
-      baseURL: this.options.baseThumbnailURL || 'https://previews.alekeagle.me'
-    });
-    return { hello: res.payload.hello, version: res.payload.version };
-  };
-
-  public getThumbnail = async (filename: string) => {
-    let res = await fetch(
-      `${
-        this.options.baseThumbnailURL || 'https://previews.alekeagle.me'
-      }/${filename}`
-    );
-    return await res.arrayBuffer();
-  };
-
-  public getSelfSessionByID = this.manufactureMethodGet<
-    [string | null],
-    __Cumulonimbus.Data.Session
-  >(sid => (sid ? `/user/session/${sid}` : '/user/session'));
-
-  public getSelfSessions = this.manufactureMethod<
-    [number, number] | [],
-    __Cumulonimbus.Data.List<__Cumulonimbus.Data.Session>
-  >(
-    (limit, offset) =>
-      `/user/sessions?limit=${limit || 50}&offset=${offset || 0}`,
-    'GET',
-    {},
-    () => null
-  );
-
-  public deleteSelfSessionByID = this.manufactureMethod<
-    [string],
-    __Cumulonimbus.Data.Session
-  >(session => `/user/session/${session}`, 'DELETE', {}, null);
-
-  public bulkDeleteSelfSessionsByID = this.manufactureMethod<
-    [string[]],
-    __Cumulonimbus.Data.DeleteBulk
-  >('/user/sessions', 'DELETE', WITH_BODY, sessions =>
-    JSON.stringify({ sessions })
-  );
-
-  public bulkDeleteAllSelfSessions = this.manufactureMethod<
-    [boolean | null],
-    __Cumulonimbus.Data.DeleteBulk
-  >(
-    allButSelf =>
-      `/user/sessions/all?allButSelf=${JSON.stringify(allButSelf || false)}`,
-    'DELETE',
-    {},
-    null
-  );
-
-  public getSelfUser = this.manufactureMethodGet<[], __Cumulonimbus.Data.User>(
-    '/user'
-  );
-
-  public editSelfUser = this.manufactureMethod<
-    [string, { username?: string; newPassword?: string; email?: string }],
-    __Cumulonimbus.Data.User
-  >('/user', 'PATCH', WITH_BODY, (password, body) => {
-    let payload: { [key: string]: string } = {};
-    Object.entries(body).forEach(a => {
-      if (a[1]) payload[a[0]] = a[1];
-    });
-    return JSON.stringify({ ...payload, password });
-  });
-
-  public deleteSelfUser = this.manufactureMethod<
-    [string, string],
-    __Cumulonimbus.Data.User
-  >('/user', 'DELETE', WITH_BODY, (username, password) =>
-    JSON.stringify({ username, password })
-  );
-
-  public editSelfDomain = this.manufactureMethod<
-    [string, string | null],
-    __Cumulonimbus.Data.User
-  >(
-    () => '/user/domain',
-    'PATCH',
-    WITH_BODY,
-    (domain, subdomain) => JSON.stringify({ domain, subdomain })
-  );
-
-  public getDomains = this.manufactureMethodGet<
-    [number | null, number | null],
-    __Cumulonimbus.Data.List<__Cumulonimbus.Data.Domain>
-  >((limit, offset) => `/domains?limit=${limit || 50}&offset=${offset || 0}`);
-
-  public getDomainsSlim = this.manufactureMethodGet<
-    [],
-    __Cumulonimbus.Data.List<__Cumulonimbus.Data.DomainSlim>
-  >('/domains/slim');
-
-  public getDomainByID = this.manufactureMethodGet<
-    [string],
-    __Cumulonimbus.Data.Domain
-  >(id => `/domain/${id}`);
-
-  public getSelfFiles = this.manufactureMethodGet<
-    [number | null, number | null],
-    __Cumulonimbus.Data.List<__Cumulonimbus.Data.File>
-  >(
-    (limit, offset) => `/user/files?limit=${limit || 50}&offset=${offset || 0}`
-  );
-
-  public getSelfFileByID = this.manufactureMethodGet<
-    [string],
-    __Cumulonimbus.Data.File
-  >(id => `/user/file/${id}`);
-
-  public deleteSelfFileByID = this.manufactureMethod<
-    [string],
-    __Cumulonimbus.Data.File
-  >(id => `/user/file/${id}`, 'DELETE', {}, null);
-
-  public bulkDeleteSelfFilesByID = this.manufactureMethod<
-    [string[]],
-    __Cumulonimbus.Data.DeleteBulk
-  >('/user/files', 'DELETE', WITH_BODY, files => JSON.stringify({ files }));
-
-  public bulkDeleteAllSelfFiles = this.manufactureMethod<
-    [],
-    __Cumulonimbus.Data.DeleteBulk
-  >('/user/files/all', 'DELETE', {}, null);
-
-  public getInstructions = this.manufactureMethodGet<
-    [number | null, number | null],
-    __Cumulonimbus.Data.List<__Cumulonimbus.Data.Instruction>
-  >(
-    (limit, offset) =>
-      `/instructions?limit=${limit || 50}&offset=${offset || 0}`
-  );
-
-  public getInstructionByID = this.manufactureMethodGet<
-    [string],
-    __Cumulonimbus.Data.Instruction
-  >(id => `/instruction/${id}`);
-
-  public getUsers = this.manufactureMethodGet<
-    [number | null, number | null],
-    __Cumulonimbus.Data.List<__Cumulonimbus.Data.User>
-  >((limit, offset) => `/users?limit=${limit || 50}&offset=${offset || 0}`);
-
-  public getUserByID = this.manufactureMethodGet<
-    [string],
-    __Cumulonimbus.Data.User
-  >(id => `/user/${id}`);
-
-  public editUserByID = this.manufactureMethod<
-    [
-      string,
-      { username?: string; password?: string; email?: string; staff?: boolean }
-    ],
-    __Cumulonimbus.Data.User
-  >(
-    id => `/user/${id}`,
-    'PATCH',
-    WITH_BODY,
-    (id, newContent) => {
-      let payload: { [key: string]: string | boolean } = {};
-      Object.entries(newContent).forEach(a => {
-        if (a[1] !== null || a[1] !== undefined) payload[a[0]] = a[1];
-      });
-      return JSON.stringify(payload);
-    }
-  );
-
-  public editUserDomain = this.manufactureMethod<
-    [string, string, string | null],
-    __Cumulonimbus.Data.User
-  >(
-    id => `/user/${id}/domain`,
-    'PATCH',
-    WITH_BODY,
-    (id, domain, subdomain) => JSON.stringify({ domain, subdomain })
-  );
-
-  public toggleUserBan = this.manufactureMethod<
-    [string],
-    __Cumulonimbus.Data.User
-  >(id => `/user/${id}/ban`, 'PATCH', {}, null);
-
-  public deleteUserByID = this.manufactureMethod<
-    [string],
-    __Cumulonimbus.Data.User
-  >(id => `/user/${id}`, 'DELETE', {}, null);
-
-  public bulkDeleteUsersByID = this.manufactureMethod<
-    [string[]],
-    __Cumulonimbus.Data.DeleteBulk
-  >(
-    () => '/users',
-    'DELETE',
-    WITH_BODY,
-    users => JSON.stringify({ users })
-  );
-
-  public createDomain = this.manufactureMethod<
-    [string, boolean],
-    __Cumulonimbus.Data.Domain
-  >('/domain', 'POST', WITH_BODY, (domain, allowsSubdomains) =>
-    JSON.stringify({ domain, allowsSubdomains })
-  );
-
-  public editDomainByID = this.manufactureMethod<
-    [string, boolean],
-    __Cumulonimbus.Data.Domain
-  >(
-    id => `/domain/${id}`,
-    'PATCH',
-    WITH_BODY,
-    (id, allowsSubdomains) => JSON.stringify({ allowsSubdomains })
-  );
-
-  public deleteDomainByID = this.manufactureMethod<
-    [string],
-    __Cumulonimbus.Data.Domain
-  >(id => `/domain/${id}`, 'DELETE', {}, null);
-
-  public bulkDeleteDomainsByID = this.manufactureMethod<
-    [string[]],
-    __Cumulonimbus.Data.DeleteBulk
-  >('/domains', 'DELETE', WITH_BODY, domains => JSON.stringify({ domains }));
-
-  public getFiles = this.manufactureMethodGet<
-    [number | null, number | null],
-    __Cumulonimbus.Data.List<__Cumulonimbus.Data.File>
-  >((limit, offset) => `/files?limit=${limit || 50}&offset=${offset || 0}`);
-
-  public getUserFiles = this.manufactureMethodGet<
-    [string, number | null, number | null],
-    __Cumulonimbus.Data.List<__Cumulonimbus.Data.File>
-  >(
-    (userID, limit, offset) =>
-      `/user/${userID}/files?limit=${limit || 50}&offset=${offset || 0}`
-  );
-
-  public getFileByID = this.manufactureMethodGet<
-    [string],
-    __Cumulonimbus.Data.File
-  >(fileID => `/file/${fileID}`);
-
-  public deleteFileByID = this.manufactureMethod<
-    [string],
-    __Cumulonimbus.Data.File
-  >(fileID => `/file/${fileID}`, 'DELETE', {}, null);
-
-  public bulkDeleteAllUserFiles = this.manufactureMethod<
-    [string],
-    __Cumulonimbus.Data.DeleteBulk
-  >(userID => `/user/${userID}/files/all`, 'DELETE', {}, null);
-
-  public bulkDeleteFilesByID = this.manufactureMethod<
-    [string[]],
-    __Cumulonimbus.Data.DeleteBulk
-  >('/files', 'DELETE', WITH_BODY, files => JSON.stringify({ files }));
-
-  public createInstruction = this.manufactureMethod<
-    [string, string[], string | null, string, string, string],
-    __Cumulonimbus.Data.Instruction
-  >(
-    '/instruction',
-    'POST',
-    WITH_BODY,
-    (name, steps, filename, fileContent, description, displayName) =>
-      JSON.stringify({
-        name,
-        steps,
-        filename,
-        fileContent,
-        description,
-        displayName
-      })
-  );
-
-  public editInstructionByID = this.manufactureMethod<
-    [
-      string,
-      {
-        steps?: string[];
-        filename?: string | null;
-        fileContent?: string;
-        description?: string;
-        displayName?: string;
       }
-    ],
-    __Cumulonimbus.Data.Instruction
-  >(
-    id => `/instruction/${id}`,
-    'PATCH',
-    WITH_BODY,
-    (id, newContent) => {
-      let payload: { [key: string]: string | string[] } = {};
-      Object.entries(newContent).forEach(a => {
-        if (a[1]) payload[a[0]] = a[1];
-      });
-      return JSON.stringify(payload);
+    );
+
+    let ratelimit: Cumulonimbus.RatelimitData | null = null;
+
+    if (res.headers.get("X-Ratelimit-Limit")) {
+      ratelimit = {
+        max: Number(res.headers.get("X-RateLimit-Limit")),
+        remaining: Number(res.headers.get("X-RateLimit-Remaining")),
+        reset: Number(res.headers.get("X-RateLimit-Reset")),
+      };
     }
-  );
 
-  public deleteInstructionByID = this.manufactureMethod<
-    [string],
-    __Cumulonimbus.Data.Instruction
-  >(id => `/instruction/${id}`, 'DELETE', {}, null);
+    const json = await res.json();
 
-  public bulkDeleteInstructionsByID = this.manufactureMethod<
-    [string[]],
-    __Cumulonimbus.Data.DeleteBulk
-  >('/instructions', 'DELETE', WITH_BODY, instructions =>
-    JSON.stringify({ instructions })
-  );
+    if (!res.ok) throw new Cumulonimbus.ResponseError(json, ratelimit);
 
-  public getUserSessionsByID = this.manufactureMethodGet<
-    [string, number | null, number | null],
-    __Cumulonimbus.Data.List<__Cumulonimbus.Data.Session>
-  >(
-    (id, limit, offset) =>
-      `/user/${id}/sessions?limit=${limit || 50}&offset=${offset || 0}`
-  );
-
-  public getUserSessionByID = this.manufactureMethodGet<
-    [string, string],
-    __Cumulonimbus.Data.Session
-  >((id, sid) => `/user/${id}/session/${sid}`);
-
-  public deleteUserSessionByID = this.manufactureMethod<
-    [string, string],
-    __Cumulonimbus.Data.Session
-  >((id, sid) => `/user/${id}/session/${sid}`, 'DELETE', {}, null);
-
-  public bulkDeleteUserSessionsByID = this.manufactureMethod<
-    [string, string[]],
-    __Cumulonimbus.Data.DeleteBulk
-  >(
-    id => `/user/${id}/sessions`,
-    'DELETE',
-    WITH_BODY,
-    (id, sessions) => JSON.stringify({ sessions })
-  );
-
-  public bulkDeleteUserSessions = this.manufactureMethod<
-    [string],
-    __Cumulonimbus.Data.DeleteBulk
-  >(id => `/user/${id}/sessions/all`, 'DELETE', {}, null);
-
-  public uploadData = this.manufactureMethod<
-    [Buffer | ArrayBuffer | Blob | File, string | undefined],
-    __Cumulonimbus.Data.SuccessfulUpload
-  >('/upload', 'POST', {}, (file, filename) => toFormData(file, filename));
+    return new Cumulonimbus(json.token, options);
+  }
 }
 
-export default {
-  Cumulonimbus: __Cumulonimbus,
-  Client
-};
+namespace Cumulonimbus {
+  export const BASE_URL = "https://alekeagle.me/api";
+  export const VERSION = version;
+
+  export interface RatelimitData {
+    max: number;
+    remaining: number;
+    reset: number;
+  }
+
+  export interface ClientOptions {
+    baseURL?: string;
+    baseThumbnailURL?: string;
+  }
+
+  export interface APICallRequestInit extends RequestInit {
+    baseURL?: string;
+  }
+
+  export interface APIResponse<T> {
+    result: T;
+    ratelimit: RatelimitData | null;
+    response: Response;
+  }
+
+  export namespace Data {
+    export interface List<T> {
+      count: number;
+      items: T[];
+    }
+
+    export interface Error {
+      code: string;
+      message: string;
+    }
+
+    export interface Success {
+      code: string;
+      message?: string;
+    }
+
+    export interface User {
+      id: string;
+      username: string;
+      email: string;
+      staff: boolean;
+      domain: string;
+      subdomain?: string;
+      bannedAt?: string;
+      createdAt: string;
+      updatedAt: string;
+    }
+
+    export interface Session {
+      iat: number;
+      exp: number;
+      sub: string;
+      name: string;
+    }
+
+    export interface Domain {
+      domain: string;
+      allowsSubdomains: boolean;
+      createdAt: string;
+      updatedAt: string;
+    }
+
+    export type DomainSlim = Omit<Domain, "createdAt" | "updatedAt">;
+
+    export interface File {
+      filename: string;
+      createdAt: string;
+      updatedAt: string;
+      userID: string;
+      size: number;
+    }
+
+    export interface DeleteBulk {
+      count: number;
+      type: "user" | "session" | "file" | "domain" | "instruction";
+    }
+
+    export interface Instruction {
+      name: string;
+      steps: string[];
+      filename: string;
+      fileContent: string;
+      description: string;
+      displayName: string;
+      createdAt: string;
+      updatedAt: string;
+    }
+
+    export interface SuccessfulAuth {
+      token: string;
+      exp: number;
+    }
+
+    export interface SuccessfulUpload {
+      url: string;
+      thumbnail: string;
+      manage: string;
+    }
+
+    export interface SanityCheck {
+      version: string;
+      hello: "world";
+    }
+  }
+
+  export interface ErrorCode {
+    INSUFFICIENT_PERMISSIONS_ERROR: "Missing Permissions";
+    INVALID_USER_ERROR: "Invalid User";
+    INVALID_PASSWORD_ERROR: "Invalid Password";
+    INVALID_SESSION_ERROR: "Invalid Session";
+    INVALID_DOMAIN_ERROR: "Invalid Domain";
+    INVALID_SUBDOMAIN_ERROR: "Invalid Subdomain: <subdomain>";
+    INVALID_FILE_ERROR: "Invalid File";
+    INVALID_INSTRUCTION_ERROR: "Invalid Instruction";
+    INVALID_ENDPOINT_ERROR: "Invalid Endpoint";
+    SUBDOMAIN_NOT_SUPPORTED_ERROR: "Subdomain Not Supported";
+    DOMAIN_EXISTS_ERROR: "Domain Exists";
+    USER_EXISTS_ERROR: "User Exists";
+    INSTRUCTION_EXISTS_ERROR: "Instruction Exists";
+    MISSING_FIELDS_ERROR: "Missing Fields: <fields>";
+    BANNED_ERROR: "Banned";
+    BODY_TOO_LARGE_ERROR: "Body Too Large";
+    RATELIMITED_ERROR: "You Have Been Ratelimited. Please Try Again Later.";
+    INTERNAL_ERROR: "Internal Server Error";
+    GENERIC_ERROR: "<message>";
+  }
+
+  // Give
+
+  export class ResponseError extends Error implements Data.Error {
+    code: keyof ErrorCode;
+    message: ErrorCode[keyof ErrorCode];
+    ratelimit: RatelimitData | null;
+    constructor(response: Data.Error, ratelimit: RatelimitData | null = null) {
+      super(response.message);
+      this.code = response.code as keyof ErrorCode;
+      this.message = response.message as ErrorCode[keyof ErrorCode];
+      this.ratelimit = ratelimit;
+    }
+  }
+}
+
+export default Cumulonimbus;
